refactor(LottoMachine): 이름 및 패키지 구조 변경
2단계를 위해 LottoMachine을 LottoController로 변경합니다.
